Issue 576. Add IO Module for Solr
Update issue 576
- type of field "threads" changed from short to int

+++ de.tudarmstadt.ukp.dkpro.core-asl/trunk/de.tudarmstadt.ukp.dkpro.core.io.solr-asl/src/main/java/de/tudarmstadt/ukp/dkpro/core/io/solr/SolrWriter_ImplBase.java
@@ -78,7 +78,7 @@ public class SolrWriter_ImplBase
      */
     public static final String PARAM_THREADS = "threads";
     @ConfigurationParameter(name = PARAM_THREADS, mandatory = true, defaultValue = "1")
-    private short threads;
+    private int threads;
 
     /**
      * When committing to the index, i.e. when all documents are processed, block until index
